requirements.txt
# Core AI Agent Framework
google-adk>=0.1.0
google-generativeai>=0.3.0

# Orbital Mechanics
skyfield>=1.46
sgp4>=2.23

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
